requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# LangChain and LLM
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.10.0
tiktoken>=0.5.2

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AWS
boto3>=1.34.0
